Backend/requirements.txt
langchain
langchain-openai
langchain-community
langserve
fastapi
faiss-cpu
sse_starlette
pydantic==1.10.13
PyPDF2
bs4
pypdf
python-slugify
langchain-chroma
langchain-chroma
lark
tiktoken
pycryptodome